requirements.txt
astroid==2.12.10
attrs==22.1.0
autopep8==1.7.0
colorama==0.4.5
dill==0.3.5.1
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
packaging==21.3
paho-mqtt==1.6.1
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycodestyle==2.9.1
pylint==2.15.3
pyparsing==3.0.9
pytest==7.1.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.5
wrapt==1.14.1
